Refactor highscore state into its own folder, and separate actions, selectors, reducer and api into it's own files
