+++ src/containers/HighscoresContainer.tsx
@@ -3,20 +3,20 @@ import { connect } from "react-redux";
 import { bindActionCreators } from 'redux';
 import HighscoresView from '../views/HighscoresView';
 import { 
-  fetchHighscoreWorld, 
-  getHighscores, 
-  getHighscoresWorld } from '../stateHandle'
+  getHighscoresWorld, 
+  getHighscoresPersonal,
+  actionGetHighscoreWorld } from '../stateHandle'
 
 const HighScoreContainer = (props) => <HighscoresView {...props} />
 
 const mapStateToProps = state => ({
-  scores: getHighscores(state),
+  scores: getHighscoresPersonal(state),
   scoresWorld: getHighscoresWorld(state)
 })
 
 const mapDispatchToProps = dispatch => {
   return bindActionCreators({
-    fetchHighscoreWorld,
+    fetchHighscoreWorld: actionGetHighscoreWorld,
   }, dispatch)
 }
