+++ src/stateHandle/highscores/index.tsx
@@ -1,0 +1,4 @@
+export * from './api'
+export * from './actions'
+export * from './reducer'
+export * from './selectors'
